requirements.txt
pycryptodome==3.23.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
flask==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
werkzeug==3.0.3
zipp==3.18.2
